Commented in timestamp functionality
Commented in the code required to limit the risk posed by the
replay-based Man-on-the-Side attack detailed in the report.

The code was not directly implemented as it is beyond the
specifications of the project and implementing it would require either
a means to estimate the maximum allowable lifespan of a message, or
altering the core functionality of the master_sign script.